Refactor(contracts): Move helper functions from entry.rs to channels.rs
